some fixes + added session.cpp
