+++ test/lab.cpp
@@ -7,7 +7,12 @@ int main(int argc, char const *argv[])
     ixm::session::arguments args;
     bool argc_check = args.argc() == argc;
     auto a1 = args[0];
-    auto a2 = args[1];
+    
+    int i = 0;
+    for (auto& a : args)
+    {
+        std::cout << i++ << ": " << a << '\n';
+    }
 
     return 0;
 }
